add alerts & set browser's back to login page problem

+++ src/Components/UserDetailsTable/UserDetailsTable.tsx
@@ -3,25 +3,41 @@ import { useDispatch, useSelector } from "react-redux";
 import { AppDispatch, RootState } from "../../redux/store/store";
 import { deleteUser, fetchAllUsers } from "../../redux/slices/usersSlice";
 import { updateOrDelete } from "../../redux/slices/updateOrDeleteSlice";
+import Swal from "sweetalert2";
 
 function UserDetailsTable(props: any) {
   const dispatch = useDispatch<AppDispatch>();
   useEffect(() => {
     console.log("Fetching users...");
-    dispatch(fetchAllUsers())
+    dispatch(fetchAllUsers());
   }, [dispatch]);
   const allUsers = useSelector((state: RootState) => state.users);
 
-   const handleUpdateOrDelete = (id: string) => {
-      dispatch(updateOrDelete(id));
-    };
+  const handleUpdateOrDelete = (id: string) => {
+    dispatch(updateOrDelete(id));
+  };
 
   const handleDelete = (id: string) => {
-    dispatch(deleteUser(id));
-  }
+    Swal.fire({
+      title: "Are you sure to delete this user: " + id + "?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteUser(id));
+      }
+    });
+  };
   return (
     <>
-      <div className="container mt-5" style={{height: "320px", maxHeight: "320px", overflowY: "scroll" }}>
+      <div
+        className="container mt-5"
+        style={{ height: "320px", maxHeight: "320px", overflowY: "scroll" }}
+      >
         <table className="table table-striped table-bordered">
           <thead className="table-dark sticky-top">
             <tr className="text-center fw-bolder">
@@ -34,24 +50,40 @@ function UserDetailsTable(props: any) {
           </thead>
           <tbody>
             {allUsers
-            .filter((user)=>props.search === "" || user.username.toLowerCase().includes(props.search.toLowerCase()))
-            .map((user, index) => (
-              <tr className="text-center fw-bold" key={index}>
-                <th scope="row">{index + 1}</th>
-                <td>{user.id}</td>
-                <td>{user.username}</td>
-                <td>{user.role}</td>
-                <td>
-                  <button  data-bs-toggle="modal"
-                      data-bs-target={props.target} className="btn btn-primary" onClick={() => {props.setShowModal()
-                      handleUpdateOrDelete(user.id)
-                      }
-
-                      }>Update</button>
-                  <button className="btn btn-danger mx-2" onClick={() => handleDelete(user.id)}>Delete</button>
-                </td>
-              </tr>
-            ))}
+              .filter(
+                (user) =>
+                  props.search === "" ||
+                  user.username
+                    .toLowerCase()
+                    .includes(props.search.toLowerCase())
+              )
+              .map((user, index) => (
+                <tr className="text-center fw-bold" key={index}>
+                  <th scope="row">{index + 1}</th>
+                  <td>{user.id}</td>
+                  <td>{user.username}</td>
+                  <td>{user.role}</td>
+                  <td>
+                    <button
+                      data-bs-toggle="modal"
+                      data-bs-target={props.target}
+                      className="btn btn-primary"
+                      onClick={() => {
+                        props.setShowModal();
+                        handleUpdateOrDelete(user.id);
+                      }}
+                    >
+                      Update
+                    </button>
+                    <button
+                      className="btn btn-danger mx-2"
+                      onClick={() => handleDelete(user.id)}
+                    >
+                      Delete
+                    </button>
+                  </td>
+                </tr>
+              ))}
           </tbody>
         </table>
       </div>
